expose print_aperture_exception() utility
