+++ tools/prometheus_ambassador/src/PrintCaughtException.cc
@@ -6,23 +6,11 @@
 
 #include "PrintCaughtException.h"
 #include <util/TypeName.h>
+#include <util/ExceptionUtil.h>
 #include <aperturedb/Exception.h>
 #include <comm/Exception.h>
 #include <sstream>
 #include <cstring>
-
-namespace {
-    template<typename EX>
-    std::string print_aperture_exception(const EX& e) {
-        std::ostringstream oss;
-        oss << type_name(e) << " " << e.name << " (" << e.file << ":" << e.line << ")";
-        if (e.errno_val != 0)
-            oss << " [code=" << strerror(e.errno_val) << "]";
-        if (!e.msg.empty())
-            oss << " " << e.msg;
-        return oss.str();
-    }
-}
 
 std::string print_caught_exception() {
     try {
